Improved Gradle releasing plugin
 - Change the plugin hierarchy so that we can apply plugins in submodules. Now the user can publish Gradle plugin from a submodule.
 - Instead of applying "com.gradle.plugin-publish" directly to root, we are letting the user to do it. Shipkit will automatically configure every submodule that uses "com.gradle.plugin-publish" plugin. This is the same strategy that we have for "java" plugin.
